% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyectosanid.R
\docType{data}
\name{proyectosanid}
\alias{proyectosanid}
\title{Histórico de proyectos adjudicados ANID}
\format{
Un data frame con 37,194 filas y 23 variables:
\itemize{
\item \code{region_ejecución}: ...
\item \code{institucion_principal}: ...
\item \code{codigo_proyecto}: ...
\item \code{n}: ...
\item \code{subdireccion}: ...
\item \code{programa}: ...
\item \code{instrumento}: ...
\item \code{nombre_concurso}: ...
\item \code{ano_curso}: ...
\item \code{ano_fallo}: ...
\item \code{nombre_proyecto}: ...
\item \code{area_ocde}: ...
\item \code{disciplina}: ...
\item \code{grupo_evaluacion}: ...
\item \code{duracion_meses}: ...
\item \code{tipo_beneficiario}: ...
\item \code{nombre_responsable}: ...
\item \code{sexo}: ...
\item \code{macrozona}: ...
\item \code{monto_adjudicado_miles}: ...
\item \code{sinfo_nosolicita}: ...
\item \code{ano_finalizacion}: ...
\item \code{ubicacion_institucion}: ...
}
}
\source{
Los datos fueron recopilados del repositorio público ANID-GITHUB/Historico-de-Proyectos-Adjudicados
usando el archivo 'BDH_Proyectos.csv'.
}
\usage{
proyectosanid
}
\description{
Datos históricos de los proyectos adjudicados bajo ANID.
}
\keyword{datasets}
